started refactoring into analysis pass
