+++ libraries/anvill_passes/src/SwitchLoweringPass.h
@@ -2,6 +2,8 @@
 
 #include <llvm/IR/Instructions.h>
 #include <llvm/Pass.h>
+#include "IndirectJumpPass.h"
+
 /*
 The goal here is to lower anvill_complete_switch to an llvm switch when we can recover the cases. This analysis must be sound but anvill_complete_switch maybe used for any complete set of indirect targets
 so cases may not even exist.
@@ -20,20 +22,21 @@ namespace anvill {
 
 
 
-    class SwitchLoweringPass: public llvm::FunctionPass{
+    class SwitchLoweringPass: public IndirectJumpPass<SwitchLoweringPass> {
         
         private:
-            static char ID;
             const std::shared_ptr<MemoryProvider> memProv;
             
         public:
-            SwitchLoweringPass(std::shared_ptr<MemoryProvider> memProv) : llvm::FunctionPass(ID), memProv(std::move(memProv)) {
+            SwitchLoweringPass(std::shared_ptr<MemoryProvider> memProv) : memProv(std::move(memProv)) {
 
             }
 
             llvm::StringRef getPassName() const override;
 
             void getAnalysisUsage(llvm::AnalysisUsage &AU) const override;
-            bool runOnFunction(llvm::Function &F) override;
+
+            bool runOnIndirectJump(llvm::CallInst* indirectJump);
+            
     };
 }
